chore: mandatory (#25)
* Feature/plan obj (#4)

* norm: sphere,plane (#6)

* add hello

* changed: double -> float, norminette: fixed.

* matrix_iverse: i->col | j->row

* mtx_inverse norminette  solution

* mtx_inverse: count[0]->c[COL], count[1]->c[ROW]

---------


* feature:parsing-and-key (#9)

* key_input: shpere, camera transformation | interface : all working good.

* norminette1 interface

* interface nominette2, added line around window, changed drawing and image size of template for solving rerender problem

* event_function: seperate, plane_key_transformation: added, push befor change t_lht -> t_obj

* change t_lht to t_obj

* seperate lights and objects

---------


* fix: panel-ui

* set panel width

* added render, added mouse hook, fix interface(first+ wiht n value, draw white at space, dynamic), added data_initailise

* norminette: fixed, optimized

---------


* Feature/object cone cylinder (#11)

* new func

* new func

* cylinder formula

* cone formula

* just fix plane

* fix: panel-ui

* set panel width

* added render, added mouse hook, fix interface(first+ wiht n value, draw white at space, dynamic), added data_initailise

* norminette: fixed, optimized

---------


* new func

* cylinder formula

* cone formula

* rebase develop

* just fix plane

---------


* norm: mandatory (#15)

* saved

* saved

* cone norm

* finish cone

* use cam

* update fov formula

* norminetted, cflags allowed

---------


* plane rotated norm

* finish plane

* fix: plane rotation (#19)

* fix: cylinder rotated (#21)

* solved (#22) cone norm rotation

* fix: add rotation objects with radian feature. (#24)

* add n_radian into t_data for rotation, added object rotatio with radian.

* open cflags

---------


---------

+++ srcs/utils/vector2.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vector_utils1.c                                    :+:      :+:    :+:   */
+/*   vector2.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: araiva <tsomsa@student.42.fr>              +#+  +:+       +#+        */
+/*   By: sharnvon <sharnvon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/05 14:11:52 by araiva            #+#    #+#             */
-/*   Updated: 2023/02/05 14:11:53 by araiva           ###   ########.fr       */
+/*   Updated: 2023/03/10 00:10:31 by sharnvon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
  * @param x value in x axis
  * @param y value in y axis
  * @param z value in z axis
- * @return t_vtr 
+ * @return t_vtr
  */
-t_vtr vtrset(float x, float y, float z)
+t_vtr	vtrset(float x, float y, float z)
 {
 	t_vtr	v;
 
@@ -34,7 +34,7 @@ t_vtr vtrset(float x, float y, float z)
 */
 float	vtrmag(t_vtr v)
 {
-	return (sqrt(pow(v.x, 2) + pow(v.y, 2) + pow(v.z, 2)));
+	return (sqrtf(pow(v.x, 2) + pow(v.y, 2) + pow(v.z, 2)));
 }
 
 /**
